Make sure jQuery's been loaded before avoiding loading it
Pull #53 is great but just a small issue: it breaks the 'share' page because `useExistingjQuery` is true then yet the share page is a standalone html page, which should use the vendored jQuery one. This checks for jQuery's existence to require loading jQuery first when applicable

I also think the 'share.html' page should be symlinked over Rakefile like the rest under `rake copy_files` task. Let me know if that's not supposed to be