tests: Add MiniProfilerConcurrencyTest
Test MiniProfiler.Step() with in multiple, nested threads or tasks and
assert the profiling report includes the correct durations for both real
time and simulated time.

The real time tests are a bit flaky and could probably use a better
approach, but to real time testing is non-deterministic and not easy to
reliably test.