Entity Framework 6 Simplification (#176)
Previously we were relying on reflection for the .Instance static access on the type. But this is realistic scenario - you need to wrap *another* provider to do anything useful (ya know, profiling). So we need not be concerned with this access pattern and can be faster and support more without it.

This should resolve #175 and be faster and simpler for everyone.