Entity Framework 6 Profiling Enhancements (#189)
* Cache the EFProfiledDbProviderServices instance.

Internally DbProviderServices is registered as a singleton. However when
the service is "replaced" via `ReplaceService<>()`, the interception func
is executed *every time* the service is resolved. We can cache the
`EFProfiledDbProviderServices` instance based on the type and key.

* Registering both a custom `DbProviderFactory` and `IDbConnectionFactory`
with EF to enable EF to use the `ProfiledDbConnection`.

This solves an issue where sometimes a `DbCommand` is created directly from
the `DbConnection` and not cached via the `DbProviderServices`.

Unfortunately to support wrapping `DbProviderFactory` and `DbConnection`,
we need to implement our own `IDbProviderFactoryResolver` and resolver for
`IProviderInvariantName`. The implementation of
`EFProfiledDbProviderFactoryResolver` in particular is
**not pretty**. But it works.
* dedicated cache per type to avoid complex key and casting