EntityFrameworkCore: remove support before netstandard2.0
The facts are that the output of events to listen to changed to drastically in 2.0 on the EntityFramework side that trying to analyze across package versions is just crazy. Given the timeline of MiniProfiler release (after 2.0 is RTM, which is already the case), it makes sense to just support EF Core 2.0 and up.

This should resolve the issues users are hitting in #209.