Options shuffling + fixes for #221
The issue in #221 centers around needing an options instance when restored from these non-referential data stores (e.g. anything not-memory). This stores he options instance as DefaultOptions statically upon configuration so that it can be used as a default for deserialized instances.

The assumption here is that the vast majority of people only want one options instance anyway, so this has basically zero overhead and covers the most common case intuitively.