Changes for #239 - don't render <script> if not authorixed.
I'm not sure why we ever did this. *Potentially* a page request isn't authroized but an AJAX request is, but this change banks on that being very low to zero percentage. Stack Overflow for example has a helper that does exactly this already. In short: why render a GUID the user can't fetch? They'll just get a deny on the AJAX request.

The old behavior rendered a needless .js include for users who couldn't load anything which is totally against the priciples of a minimally impactful profiler.