ASP.NET Core 3: use System.Text.Json for async streams
Synchronous stream reads are disabled by default in ASP.NET Core 3.x, and we wouldn't want to be a bottleneck in any case. This uses the System.Text.Json package instead (a bit hackily inline since one path is async and the old isn't) but only in netcoreapp3.0 (while it's in preview). Unfortunately, System.Text.Json doesn't support HTML encoded output directly, so I'll need to see about the serialization replacement of Newtonsoft in any stable release. Hopefully we're not pulling in 2 serializers forever due to that 1 detail.

This fixes a bug where multiple profilers were showing because the last one was fetched repeatedly, due to failing to parse an Id with a synchronous stream read.