Default to 2 decimal places for timings and add options to control it (#580)
This adds a new `PopupDecimalPlaces` option to control, but defaults to 2 now that things are getting so fast we need more precision. For anyone who finds this cluttered options to control it have been added and the old behavior is 1 setting away. Added to samples and examples.

Example from sample:
![example](https://user-images.githubusercontent.com/454813/138612824-09450896-080a-49ef-bc91-71f1ef031aea.png)