Optimize MVC diagnostic listener (#584)
Meant to do this a while ago - stop generating strings for repeated things. We have a finite amount of combinations here, so let's use that to our advantage.

Overall, we most often have a pair of one label that doesn't change at all and a second string within a finite set. Because of this, we can cache the tuple or 2 => string conversion and not allocate to gen 0 per step. Since the conversion of source => final string on the second half only happens once, we can also prettify it a bit without paying the cost each time.